Use Strich by default
