+++ apps/passport-client/components/screens/MultiChoiceScanScreen.tsx
@@ -1,6 +1,5 @@
 import { useCallback, useState } from "react";
 import { toast } from "react-hot-toast";
-import { QrReader } from "react-qr-reader";
 import { useNavigate } from "react-router-dom";
 import { SingleValue } from "react-select";
 import styled from "styled-components";
@@ -8,6 +7,7 @@ import { useLaserScannerKeystrokeInput } from "../../src/appHooks";
 import { loadUsingLaserScanner } from "../../src/localstorage";
 import { maybeRedirect } from "../../src/util";
 import { H5, Spacer, TextCenter } from "../core";
+import { ReactQrReaderScanner } from "../core/scanners/ReactQRReaderScanner";
 import { ScanditScanner } from "../core/scanners/ScanditScanner";
 import { StrichScanner } from "../core/scanners/StrichScanner";
 import { AppContainer } from "../shared/AppContainer";
@@ -17,19 +17,6 @@ import {
   Back,
   Home
 } from "./ScannedTicketScreens/PodboxScannedTicketScreen/PodboxScannedTicketScreen";
-
-const ButtonsContainer = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: row;
-  gap: 8px;
-  margin-bottom: 16px;
-
-  button {
-    flex-grow: 1;
-  }
-`;
 
 // Scan a PCD QR code, then go to /verify to verify and display the proof.
 export function MultiChoiceScanScreen(): JSX.Element {
@@ -81,17 +68,6 @@ export function MultiChoiceScanScreen(): JSX.Element {
     [nav]
   );
 
-  const onQrReaderResult = useCallback(
-    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-    // @ts-ignore
-    (result) => {
-      if (result) {
-        onResult(result.getText());
-      }
-    },
-    [onResult]
-  );
-
   return (
     <AppContainer bg="gray">
       {!usingLaserScanner && (
@@ -108,13 +84,7 @@ export function MultiChoiceScanScreen(): JSX.Element {
           ></Select>
           <Spacer h={16} />
           {scanner === "react-qr-reader" && (
-            <QrReader
-              className="qr"
-              onResult={onQrReaderResult}
-              constraints={{ facingMode: "environment", aspectRatio: 1 }}
-              ViewFinder={ViewFinder}
-              containerStyle={{ width: "100%" }}
-            />
+            <ReactQrReaderScanner onResult={onResult} />
           )}
           {scanner === "scandit" && <ScanditScanner onScan={onResult} />}
           {scanner === "strich" && <StrichScanner onResult={onResult} />}
@@ -153,64 +123,28 @@ export function MultiChoiceScanScreen(): JSX.Element {
   );
 }
 
-function ViewFinder(): JSX.Element {
-  return (
-    <ScanOverlayWrap>
-      <Guidebox>
-        <Corner top left />
-        <Corner top />
-        <Corner left />
-        <Corner />
-      </Guidebox>
-    </ScanOverlayWrap>
-  );
-}
-
 const Orange = styled.span`
   font-weight: bold;
   color: orange;
-`;
-
-const ScanOverlayWrap = styled.div`
-  position: absolute;
-  top: 0;
-  left: 0;
-  bottom: 0;
-  right: 0;
-  z-index: 1;
-  margin: 16px;
 `;
 
 const FullWidthRow = styled.div`
   width: 100%;
 `;
 
-const Guidebox = styled.div`
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  width: 75%;
-  height: 75%;
-  background: rgba(255, 255, 255, 0.05);
-  border-radius: 8px;
-`;
-
-const Corner = styled.div<{ top?: boolean; left?: boolean }>`
-  position: absolute;
-  ${(p): string => (p.top ? "top: 0" : "bottom: 0")};
-  ${(p): string => (p.left ? "left: 0" : "right: 0")};
-  border: 2px solid white;
-  ${(p): string => (p.left ? "border-right: none" : "border-left: none")};
-  ${(p): string => (p.top ? "border-bottom: none" : "border-top: none")};
-  width: 16px;
-  height: 16px;
-  ${(p): string => (p.left && p.top ? "border-radius: 8px 0 0 0;" : "")};
-  ${(p): string => (p.left && !p.top ? "border-radius: 0 0 0 8px;" : "")};
-  ${(p): string => (!p.left && p.top ? "border-radius: 0 8px 0 0;" : "")};
-  ${(p): string => (!p.left && !p.top ? "border-radius: 0 0 8px 0;" : "")};
-`;
-
 const QRContainer = styled.div`
   width: 100%;
+`;
+
+const ButtonsContainer = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: row;
+  gap: 8px;
+  margin-bottom: 16px;
+
+  button {
+    flex-grow: 1;
+  }
 `;
