+++ apps/passport-client/src/dispatch.ts
@@ -28,6 +28,7 @@ import {
   SemaphoreIdentityPCDTypeName
 } from "@pcd/semaphore-identity-pcd";
 import { assertUnreachable, sleep } from "@pcd/util";
+import { StrichSDK } from "@pixelverse/strichjs-sdk";
 import { Identity } from "@semaphore-protocol/identity";
 import _ from "lodash";
 import { createContext } from "react";
@@ -151,6 +152,9 @@ export type Action =
       type: "merge-import";
       collection: PCDCollection;
       pcdsToMergeIds: Set<PCD["id"]>;
+    }
+  | {
+      type: "initialize-strich";
     };
 
 export type StateContextValue = {
@@ -264,6 +268,8 @@ export async function dispatch(
         action.collection,
         action.pcdsToMergeIds
       );
+    case "initialize-strich":
+      return initializeStrich(state, update);
     default:
       // We can ensure that we never get here using the type system
       return assertUnreachable(action);
@@ -1247,4 +1253,19 @@ async function removeAllPCDsInFolder(
   await savePCDs(state.pcds);
   update({ pcds: state.pcds });
   window.scrollTo({ top: 0 });
+}
+
+async function initializeStrich(
+  state: AppState,
+  update: ZuUpdate
+): Promise<void> {
+  if (!appConfig.strichLicenseKey) {
+    throw new Error("Strich license key is not defined");
+  }
+  try {
+    await StrichSDK.initialize(appConfig.strichLicenseKey);
+    update({ strichSDKstate: "initialized" });
+  } catch (e) {
+    update({ strichSDKstate: "error" });
+  }
 }
