+++ apps/passport-client/src/loadInitialState.ts
@@ -61,7 +61,8 @@ export async function loadInitialState(): Promise<AppState> {
     offline: !window.navigator.onLine,
     serverStorageRevision: persistentSyncStatus.serverStorageRevision,
     serverStorageHash: persistentSyncStatus.serverStorageHash,
-    importScreen: undefined
+    importScreen: undefined,
+    strichSDKstate: undefined
   };
 
   if (!validateAndLogInitialAppState("loadInitialState", state)) {
